pyproject.toml
[tool.poetry]
name = "whisper-worker"
version = "0.1.0"
description = "Transcription worker for Whisper"
authors = ["Alexander Ditzend <alex@saia.ar>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
whisper = {git = "https://github.com/openai/whisper.git", rev="b9265e5796f5d80c18d1f9231ab234225676780b"}
python-dotenv = "^0.21.1"
coloredlogs = "^15.0.1"
pika = "^1.3.1"
ffmpeg-python = "^0.2.0"
transformers = "^4.22.1"
tqdm = "^4.64.1"
unidecode = "^1.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
